// app/models/character.type.ts
export interface Character {
    url: string;
    name: string;
    culture: string;
    born: string;
    died: string;
    titles: any[];
    aliases: string[];
    father: string;
    mother: string;
    spouse: string;
    allegiances: any[];
    books: string[];
    povBooks: any[];
    tvSeries: any[];
    playedBy: any[];
}